chore(deps-dev): bump `vitepress-plugin-group-icons` from `1.3.6` to `1.3.7` (#265)
This pull request includes minor updates to the Korean documentation and
a version bump for a plugin in the `websites/vitepress` package.

Documentation update:

*
[`websites/vitepress/ko/learn/how-to-parse-input-value.md`](diffhunk://#diff-23393c5e02c45691527083acfcadef7dc125af1cfcad10f20412864a0a83fbfdL275-R275):
Improved the code example by simplifying the `map` function call to
directly use `Number` instead of a lambda function.

Dependency update:

*
[`websites/vitepress/package.json`](diffhunk://#diff-10ef64608fffd18cd9a3d2a3b2adb3361c324abf4a0040a210720ab259c6acd2L19-R19):
Updated the version of `vitepress-plugin-group-icons` from `^1.3.6` to
`^1.3.7`.