fix(eslint-config-bananass): update `no-missing-import` rule to ignore type imports and add `eslint.config.mjs` to `sync-client.yml` (#270)
This pull request includes updates to configuration files and
adjustments to ESLint rules. The most important changes are the addition
of a new configuration file and an update to an existing ESLint rule.

Configuration updates:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R73-R74):
Added `eslint.config.mjs` to the list of files to be synced.

ESLint rule adjustments:

*
[`packages/eslint-config-bananass/src/rules/node/node.js`](diffhunk://#diff-c96c91af708b289359a5b68fb41a21628063ede30d509cc8901b9063cacd9309L110-R115):
Modified the `n/no-missing-import` rule to ignore type imports by adding
an options object.