ci(*): create `release.yml` (#277)
This pull request includes changes to the GitHub Actions workflows to
add a new release workflow and update the sync-client configuration. The
most important changes are as follows:

### Addition of release workflow:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R1-R32):
Added a new workflow to handle releases when a pull request is merged,
ensuring it only triggers under specific conditions.

### Update to sync-client configuration:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R18-R19):
Updated the sync-client configuration to include the new release
workflow.