chore(*): update `tsconfig.json` to include patterns to support `.mjs` and `.cjs` files (#285)
And also fix a heading in `CHANGELOG.md`

---

This pull request includes changes to the `CHANGELOG.md` file and
updates to the TypeScript configuration files across multiple packages
to support additional JavaScript module formats.

Changes to documentation:

*
[`CHANGELOG.md`](diffhunk://#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4edL1-R1):
Updated the title from "bananass" to "Change Log" and added links to
Releases and Tags on GitHub.

Updates to TypeScript configuration:

*
[`packages/bananass-utils-console/tsconfig.json`](diffhunk://#diff-00d843065b36c05dd508bb0a94f06c16cc20bb2c4a760080908e336b719d3232L14-R14):
Included support for `.mjs` and `.cjs` module formats in addition to
`.js` files.
*
[`packages/bananass-utils-vitepress/tsconfig.json`](diffhunk://#diff-00d843065b36c05dd508bb0a94f06c16cc20bb2c4a760080908e336b719d3232L14-R14):
Included support for `.mjs` and `.cjs` module formats in addition to
`.js` files.
*
[`packages/bananass/tsconfig.json`](diffhunk://#diff-00d843065b36c05dd508bb0a94f06c16cc20bb2c4a760080908e336b719d3232L14-R14):
Included support for `.mjs` and `.cjs` module formats in addition to
`.js` files.
*
[`packages/create-bananass/tsconfig.json`](diffhunk://#diff-00d843065b36c05dd508bb0a94f06c16cc20bb2c4a760080908e336b719d3232L14-R14):
Included support for `.mjs` and `.cjs` module formats in addition to
`.js` files.