fix(eslint-config-bananass): add `.d.ts`, `.d.mts` and `.d.cts` to `node.resolverConfig.extensions` (#298)
This pull request includes a few changes to improve the codebase and
configuration settings. The most important changes are the addition of a
TODO comment for future asynchronous implementation and updating the
list of file extensions in the ESLint configuration.

Improvements and configuration updates:

*
[`packages/bananass/src/core/fs/find-root-dir/find-root-dir.js`](diffhunk://#diff-aed2769c5bf3d27e1aa5d4b258ebffc32608a930d0f9f6a9d4c31e6416de2122R5-R6):
Added a TODO comment to indicate the need for future asynchronous
implementation.
*
[`packages/eslint-config-bananass/src/settings.js`](diffhunk://#diff-f3313674a9f29402cc0fbfc6fa06d9c88b5b8caa48ac94dca0b71f547b85514fL12-R25):
Updated the `extensions` array to include additional TypeScript
declaration file extensions (`.d.ts`, `.d.mts`, `.d.cts`).