fix(eslint-config-bananass)!: update rule configurations for better linting accuracy (#318)
This pull request includes updates to the ESLint configuration rules in
the `packages/eslint-config-bananass` package. The changes focus on
modifying rule settings for better code quality and consistency.

### ESLint rule updates:

*
[`packages/eslint-config-bananass/src/rules/eslint/eslint-suggestions.js`](diffhunk://#diff-1f441d9b82dbdd31cd2b0b9d9d6c30b2d66b91dfed4e1f4703da5ef1e295430cL1051-R1051):
Changed the `avoidQuotes` option to `false` in the `quote-props` rule to
enforce consistent property quoting.
*
[`packages/eslint-config-bananass/src/rules/eslint/eslint-suggestions.js`](diffhunk://#diff-1f441d9b82dbdd31cd2b0b9d9d6c30b2d66b91dfed4e1f4703da5ef1e295430cL1226-R1226):
Updated the `require-unicode-regexp` rule to `'warn'` with a `{
requireFlag: 'u' }` option to encourage the use of Unicode flag in
regular expressions.

### React rule updates:

*
[`packages/eslint-config-bananass/src/rules/react/react.js`](diffhunk://#diff-d2e14b50ea9ff3e279f30d972b03fd16efe19993bd0d6670de8571d2213515c5L816-R820):
Changed the `react/react-in-jsx-scope` rule from `'warn'` to `'error'`
to enforce the presence of React when using JSX.