docs(*): update `README.md`, `README.en.md`, and `CONTRIBUTING.md` (#336)
This pull request includes updates to the documentation files, primarily
focusing on improving clarity and consistency in the `CONTRIBUTING.md`,
`README.en.md`, and `README.md` files. The changes also include the
addition of badges to the README files for better visibility of the
project's status.

Documentation improvements:

*
[`CONTRIBUTING.md`](diffhunk://#diff-eca12c0a30e25b4b46522ebf89465a03ba72a03f540796c979137931d8f92055L5-R74):
Simplified the introduction and rephrased several sections for better
clarity. Updated the contribution guidelines to be more concise and
added instructions for forking, cloning, and setting up the project.

Enhancements to README files:

*
[`README.en.md`](diffhunk://#diff-b4f47ddd2e9baf4f63a8b1c7adb79cc6caabfe2958e82d73d7b80d1468075213L1-R15):
Added badges for linting, testing, code coverage, and Node.js version.
Improved the introduction and detailed the benefits of using the
Bananass framework. Updated the "Getting Started" section to use `npm
create bananass@latest` for installation.
[[1]](diffhunk://#diff-b4f47ddd2e9baf4f63a8b1c7adb79cc6caabfe2958e82d73d7b80d1468075213L1-R15)
[[2]](diffhunk://#diff-b4f47ddd2e9baf4f63a8b1c7adb79cc6caabfe2958e82d73d7b80d1468075213L24-R49)
[[3]](diffhunk://#diff-b4f47ddd2e9baf4f63a8b1c7adb79cc6caabfe2958e82d73d7b80d1468075213L59-R81)
[[4]](diffhunk://#diff-b4f47ddd2e9baf4f63a8b1c7adb79cc6caabfe2958e82d73d7b80d1468075213L79-R134)
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R15):
Similar updates as the English README, including the addition of badges
and improved descriptions. Updated the installation instructions and
emphasized the framework's support for modern JavaScript and TypeScript
features.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R15)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L24-R49)
[[3]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L59-R81)
[[4]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L79-R136)
[[5]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L144-R161)