ci(*): create `test-cross-platform.yml` (#363)
This pull request introduces a new GitHub Actions workflow to test the
codebase across multiple platforms and Node.js versions. The workflow
ensures compatibility and reliability by running tests on macOS, Ubuntu,
and Windows with various Node.js versions.

### New GitHub Actions Workflow:
*
[`.github/workflows/test-cross-platform.yml`](diffhunk://#diff-290907d9a4bcf13181105b134bd1645da5a3ef27c54f148ad5823e6b77e88426R1-R48):
Added a `test-cross-platform` workflow that triggers on pushes and pull
requests to the `main` branch. It runs tests on `macos-latest`,
`ubuntu-latest`, and `windows-latest` with Node.js versions `23.0.0`,
`22.3.0`, and `20.18.0`. The workflow includes steps for checkout,
Node.js setup, dependency installation, building (if applicable), and
running tests with coverage.