docs(websites-vitepress): create Q & A (#372)
This pull request introduces a new "Q & A" section to the documentation
and website, replacing the old `docs/question-and-answer.md` file with
localized versions in both English and Korean. It also updates
navigation menus to include links to the new pages.

### Documentation Updates:
* Deleted the outdated `docs/question-and-answer.md` file, which
contained Q&A content in Korean.
* Added a new English Q&A page
(`websites/vitepress/en/learn/q-and-a.md`) with a placeholder message
directing users to the Korean version.
* Added a new Korean Q&A page (`websites/vitepress/ko/learn/q-and-a.md`)
with updated content, including details about Node.js language
restrictions on Baekjoon and line feed usage in the Banana framework.

### Navigation Updates:
* Updated the English navigation menu
(`websites/vitepress/.vitepress/config/en.js`) to include a link to the
new Q&A page.
* Updated the Korean navigation menu
(`websites/vitepress/.vitepress/config/ko.js`) to include a link to the
new Q&A page.