chore(*): remove `is-interactive` dependency and update Codecov configuration (#441)
This pull request introduces changes to integrate Codecov for test
coverage reporting and remove the `is-interactive` dependency. The most
important changes include adding the `CODECOV_TOKEN` environment
variable to the GitHub Actions workflow, updating the VitePress
configuration to use this token for bundle analysis, and cleaning up
unused dependencies.

### Codecov integration:

*
[`.github/workflows/test.yml`](diffhunk://#diff-faff1af3d8ff408964a57b2e475f69a6b7c7b71c9978cccc8f471798caac2c88R21-R23):
Added the `CODECOV_TOKEN` environment variable to the test job in the
GitHub Actions workflow to enable Codecov integration.
*
[`websites/vitepress/.vitepress/config/shared.js`](diffhunk://#diff-73876f6a978dec27dfeeaa8c70437d5664ae37af70af0cda625a8d54373d237bL110-R109):
Updated the `codecovVitePlugin` configuration to enable bundle analysis
only when the `CODECOV_TOKEN` environment variable is set.

### Dependency cleanup:

*
[`websites/vitepress/.vitepress/config/shared.js`](diffhunk://#diff-73876f6a978dec27dfeeaa8c70437d5664ae37af70af0cda625a8d54373d237bL17):
Removed the unused `is-interactive` import.
*
[`websites/vitepress/package.json`](diffhunk://#diff-10ef64608fffd18cd9a3d2a3b2adb3361c324abf4a0040a210720ab259c6acd2L15):
Removed the `is-interactive` dependency since it is no longer used.