chore(*): update ESLint config and dependabot (#479)
This pull request includes updates to the Dependabot configuration and a
refactor in the ESLint configuration file to improve maintainability and
consistency. The most important changes are grouped below:

### Dependabot Configuration Updates:
* Reduced the `open-pull-requests-limit` from 3 to 2 for both the
`typescript-eslint` and the scoped dependencies in the
`.github/dependabot.yml` file. This helps manage the number of active
pull requests more effectively.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L33-R33)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L51-R51)

### ESLint Configuration Refactor:
* Replaced the inline `ignores` configuration with the `globalIgnores`
helper function in `eslint.config.mjs`, simplifying the configuration
and improving readability.