docs(*): remove `PULL_REQUEST_TEMPLATE.md` in favor of centralized template (#588)
This pull request removes the pull request template and updates the
synchronization configuration to reflect this change. The removal
simplifies the repository by eliminating unused or redundant files.

### Removal of pull request template:

*
[`.github/PULL_REQUEST_TEMPLATE.md`](diffhunk://#diff-18813c86948efc57e661623d7ba48ff94325c9b5421ec9177f724922dd553a35L1-L40):
Deleted the pull request template file, which previously provided a
structured format for submitting pull requests.

### Update to synchronization configuration:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L29-L30):
Removed the entry for syncing the deleted `PULL_REQUEST_TEMPLATE.md`
file to the `configs` directory.