fix(eslint-config-bananass): update ESLint configuration (#827)
This pull request updates the ESLint configuration to better align with
modern React and JavaScript best practices. The main focus is on
removing outdated React rules related to the JSX transform and
tightening restrictions on the use of browser alerts and console
statements.

**React JSX Transform Updates:**

* Disabled the `react/react-in-jsx-scope` rule in `react.js`,
`jsx-next.js`, and `tsx-next.js`, reflecting that importing React is no
longer required for JSX since React 17.
[[1]](diffhunk://#diff-d2e14b50ea9ff3e279f30d972b03fd16efe19993bd0d6670de8571d2213515c5L816-R821)
[[2]](diffhunk://#diff-ed37bd3604035f4eb1410a22dcd8e5f8a07d6d304b78c9a59425b8ccef30703cL67-L68)
[[3]](diffhunk://#diff-7e6986b8aa252958b9da97c71d62e00f1895da50f896ceb71305ba34671b403aL72-L73)
* Turned off the `react/jsx-uses-react` rule in `react.js`, as React no
longer needs to be imported for JSX usage.

**Stricter Linting for Browser APIs:**

* Changed the `no-alert` rule from `'warn'` to `'error'` in `eslint.js`,
making the use of browser alert dialogs a linting error.
* Changed the `no-console` rule from `'warn'` to `'error'` in
`eslint.js`, making the use of `console` statements a linting error.